Query for most recent of data  in database then loop to the current year to ingest newly released cot data.

+++ src/ingestion/fetch.rs
@@ -11,7 +11,7 @@ use zip::DateTime;
 // COT Data Fetching
 //=======================================================================================================
 
-pub async fn fetch_cot_data(year: i64, pool: PgPool) -> Result<(), Box<dyn Error>> {
+pub async fn fetch_cot_data(year: i32, pool: PgPool) -> Result<(), Box<dyn Error>> {
     
     let cot_request = reqwest::get(format!("https://www.cftc.gov/files/dea/history/fut_fin_txt_{}.zip", year))
         .await?
